+++ lib/docusign_rest/client.rb
@@ -7,6 +7,7 @@ module DocusignRest
     # Define the same set of accessors as the DocusignRest module
     attr_accessor *Configuration::VALID_CONFIG_KEYS
     attr_accessor :docusign_authentication_headers, :acct_id
+    attr_accessor :previous_call_log
 
     def initialize(options={})
       # Merge the config values from the module and those passed to the client.
@@ -38,6 +39,9 @@ module DocusignRest
       # the instance var @account_id because that'll override the attr_accessor
       # that is automatically configured for the configure block
       @acct_id = account_id
+
+      #initialize the log cache
+      @previous_call_log = []
     end
 
 
@@ -143,6 +147,7 @@ module DocusignRest
 
       http = initialize_net_http_ssl(uri)
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -169,7 +174,9 @@ module DocusignRest
       uri = build_uri('/login_information')
       request = Net::HTTP::Get.new(uri.request_uri, headers(options[:headers]))
       http = initialize_net_http_ssl(uri)
-      http.request(request)
+      response = http.request(request)
+      generate_log(request, response, uri)
+      response
     end
 
 
@@ -723,9 +730,9 @@ module DocusignRest
                 )
 
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
-
 
     # Public: allows a template to be dynamically created with several options.
     #
@@ -783,6 +790,7 @@ module DocusignRest
                 )
 
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -797,6 +805,7 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Get.new(uri.request_uri, headers(content_type))
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -847,6 +856,7 @@ module DocusignRest
       request.body = post_body
 
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -901,6 +911,7 @@ module DocusignRest
                 )
 
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -923,6 +934,7 @@ module DocusignRest
       request = Net::HTTP::Post.new(uri.request_uri, headers(content_type))
 
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -958,6 +970,7 @@ module DocusignRest
       request.body = post_body
 
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -982,6 +995,7 @@ module DocusignRest
       request.body = { returnUrl: options[:return_url] }.to_json
 
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1008,7 +1022,7 @@ module DocusignRest
       request.body = post_body
 
       response = http.request(request)
-
+      generate_log(request, response, uri)
       parsed_response = JSON.parse(response.body)
       parsed_response['url']
     end
@@ -1036,6 +1050,7 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Get.new(uri.request_uri, headers(content_type))
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1052,6 +1067,7 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Get.new(uri.request_uri, headers(content_type))
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1082,7 +1098,7 @@ module DocusignRest
       http     = initialize_net_http_ssl(uri)
       request  = Net::HTTP::Get.new(uri.request_uri, headers(content_type))
       response = http.request(request)
-
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1115,6 +1131,7 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Get.new(uri.request_uri, headers(content_type))
       response = http.request(request)
+      generate_log(request, response, uri)
       return response.body if options[:return_stream]
 
       split_path = options[:local_save_path].split('/')
@@ -1142,7 +1159,7 @@ module DocusignRest
       http     = initialize_net_http_ssl(uri)
       request  = Net::HTTP::Get.new(uri.request_uri, headers(content_type))
       response = http.request(request)
-
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1176,6 +1193,7 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Get.new(uri.request_uri, headers(content_type))
       response = http.request(request)
+      generate_log(request, response, uri)
       return response.body if options[:return_stream]
 
       split_path = options[:local_save_path].split('/')
@@ -1218,7 +1236,7 @@ module DocusignRest
       request = Net::HTTP::Put.new(uri.request_uri, headers(content_type))
       request.body = post_body
       response = http.request(request)
-
+      generate_log(request, response, uri)
       response
     end
 
@@ -1242,7 +1260,7 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Get.new(uri.request_uri, headers(content_type))
       response = http.request(request)
-
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1256,6 +1274,7 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Get.new(uri.request_uri, headers(content_type))
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1297,6 +1316,7 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Get.new(uri.request_uri, headers(content_type))
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1314,6 +1334,7 @@ module DocusignRest
       request = Net::HTTP::Post.new(uri.request_uri, headers(content_type))
       request.body = post_body
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1341,6 +1362,7 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Delete.new(uri.request_uri, headers(content_type))
       response = http.request(request)
+      generate_log(request, response, uri)
       json = response.body
       json = '{}' if json.nil? || json == ''
       JSON.parse(json)
@@ -1359,7 +1381,9 @@ module DocusignRest
 
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Get.new(uri.request_uri, headers({ 'Content-Type' => 'application/json' }))
-      JSON.parse(http.request(request).body)
+      response = http.request(request)
+      generate_log(request, response, uri)
+      JSON.parse(response.body)
     end
 
 
@@ -1373,7 +1397,9 @@ module DocusignRest
 
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Get.new(uri.request_uri, headers({ 'Content-Type' => 'application/json' }))
-      JSON.parse(http.request(request).body)
+      response = http.request(request)
+      generate_log(request, response, uri)
+      JSON.parse(response.body)
     end
 
 
@@ -1389,6 +1415,7 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Get.new(uri.request_uri, headers(content_type))
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1415,6 +1442,7 @@ module DocusignRest
       request.body = post_body
 
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1439,7 +1467,9 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Put.new(uri.request_uri, headers(content_type))
       request.body = post_body
-      http.request(request)
+      response = http.request(request)
+      generate_log(request, response, uri)
+      response
     end
 
     # Public deletes a document for a given envelope
@@ -1468,6 +1498,7 @@ module DocusignRest
       request.body = post_body
 
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1498,8 +1529,9 @@ module DocusignRest
       http = initialize_net_http_ssl(uri)
       request = Net::HTTP::Put.new(uri.request_uri, headers(headers))
       request.body = post_body
-
-      http.request(request)
+      response = http.request(request)
+      generate_log(request, response, uri)
+      response
     end
 
     # Public adds signers to a given envelope
@@ -1524,6 +1556,7 @@ module DocusignRest
       request.body = post_body
 
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
     end
 
@@ -1574,7 +1607,32 @@ module DocusignRest
       request.body = post_body
 
       response = http.request(request)
+      generate_log(request, response, uri)
       JSON.parse(response.body)
+    end
+
+    private
+
+    # Private: Generates a standardized log of the request and response pair
+    # to and from DocuSign for logging and API Certification.
+    # and resulting list is set to the publicly accessible: @previous_call_log
+    # For example:
+    # envelope = connection.create_envelope_from_document(doc)
+    # connection.previous_call_log.each {|line| logger.debug line }
+    def generate_log(request, response, uri)
+      log = ['--DocuSign REQUEST--']
+      log << "#{request.method} #{uri.to_s}"
+      request.each_capitalized{ |k,v| log << "#{k}: #{v.gsub(/(?<="Password":")(.+?)(?=")/, '[FILTERED]')}" }
+      # Trims out the actual binary file to reduce log size
+      if request.body
+        request_body = request.body.gsub(/(?<=Content-Transfer-Encoding: binary).+?(?=-------------RubyMultipartPost)/m, "\n[BINARY BLOB]\n")
+        log << "Body: #{request_body}"
+      end
+      log << '--DocuSign RESPONSE--'
+      log << "HTTP/#{response.http_version} #{response.code} #{response.msg}"
+      response.each_capitalized{ |k,v| log << "#{k}: #{v}" }
+      log << "Body: #{response.body}"
+      @previous_call_log = log
     end
   end
 end